chore: re-organize router core module
