+++ packages/mobile/lib/routers/enums/onboarding/app-setup-route.enum.ts
