+++ packages/mobile/lib/routers/enums/onboarding/index.ts
