+++ packages/mobile/lib/routers/enums/onboarding/network-setup-route.enum.ts
