+++ packages/mobile/lib/routers/enums/onboarding/onboarding-route.enum.ts
