+++ packages/mobile/lib/routers/enums/onboarding/storage-protection-setup-route.enum.ts
