+++ packages/mobile/lib/routers/enums/onboarding/stronghold-setup-route.enum.ts
