+++ packages/mobile/lib/routers/subrouters/onboarding/app-setup-router.ts
