+++ packages/mobile/lib/routers/subrouters/onboarding/stronghold-setup-router.ts
