+++ packages/shared/lib/core/router/app-router.ts
@@ -2,10 +2,10 @@ import { get, writable } from 'svelte/store'
 
 import { profiles } from '@core/profile'
 
+import { Router } from './classes'
 import { AppRoute, LoginRoute } from './enums'
-import { Router } from './router'
-import { FireflyEvent } from './types'
 import { loginRoute } from './subrouters'
+import { FireflyEvent } from './types'
 
 export const appRoute = writable<AppRoute>(null)
 export const appRouter = writable<AppRouter>(null)
