+++ packages/shared/lib/core/router/classes/index.ts
@@ -1,0 +1,2 @@
+export * from './router.class'
+export * from './subrouter.class'
