+++ packages/shared/lib/core/router/dashboard-router.ts
@@ -1,7 +1,7 @@
 import { writable } from 'svelte/store'
 
+import { Router } from './classes'
 import { DashboardRoute } from './enums'
-import { Router } from './router'
 
 export const dashboardRouter = writable<DashboardRouter>(null)
 export const dashboardRoute = writable<DashboardRoute>(null)
