+++ packages/shared/lib/core/router/index.ts
@@ -1,3 +1,4 @@
+export * from './classes'
 export * from './enums'
 export * from './interfaces'
 export * from './subrouters'
