+++ packages/shared/lib/core/router/interfaces/router.interface.ts
@@ -4,5 +4,5 @@ export interface IRouter {
     next(event?: FireflyEvent): void
     previous(): void
 
-    reset(): void
+    reset?(): void
 }
