+++ packages/shared/lib/core/router/subrouters/index.ts
@@ -1,4 +1,3 @@
 export * from './onboarding'
 
 export * from './login-router'
-export * from './subrouter'
