+++ packages/shared/lib/core/router/subrouters/onboarding/app-setup-router.ts
@@ -2,7 +2,7 @@ import { get, writable } from 'svelte/store'
 
 import { AppSetupRoute } from '../../enums'
 import { onboardingRouter } from '../../onboarding-router'
-import { Subrouter } from '../subrouter'
+import { Subrouter } from '../../classes'
 
 export const appSetupRoute = writable<AppSetupRoute>(null)
 export const appSetupRouter = writable<AppSetupRouter>(null)
