+++ packages/shared/lib/core/router/subrouters/onboarding/profile-backup-router.ts
@@ -2,7 +2,7 @@ import { get, writable } from 'svelte/store'
 
 import { ProfileBackupRoute } from '../../enums'
 import { onboardingRouter } from '../../onboarding-router'
-import { Subrouter } from '../subrouter'
+import { Subrouter } from '../../classes'
 
 export const profileBackupRoute = writable<ProfileBackupRoute>(null)
 export const profileBackupRouter = writable<ProfileBackupRouter>(null)
