+++ packages/shared/lib/core/router/subrouters/onboarding/storage-protection-setup-router.ts
@@ -4,7 +4,7 @@ import { onboardingProfile, ProfileProtectionType } from '@contexts/onboarding'
 
 import { onboardingRouter } from '../../onboarding-router'
 import { StorageProtectionSetupRoute } from '../../enums'
-import { Subrouter } from '../subrouter'
+import { Subrouter } from '../../classes'
 
 export const storageProtectionSetupRoute = writable<StorageProtectionSetupRoute>(null)
 export const storageProtectionSetupRouter = writable<StorageProtectionSetupRouter>(null)
