+++ packages/shared/lib/core/router/subrouters/onboarding/stronghold-setup-router.ts
@@ -2,7 +2,7 @@ import { get, writable } from 'svelte/store'
 
 import { StrongholdSetupRoute } from '../../enums'
 import { onboardingRouter } from '../../onboarding-router'
-import { Subrouter } from '../subrouter'
+import { Subrouter } from '../../classes'
 
 export const strongholdSetupRoute = writable<StrongholdSetupRoute>(null)
 export const strongholdSetupRouter = writable<StrongholdSetupRouter>(null)
